requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==1.2.2
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
cytoolz==0.11.2
dbus-python==1.2.16
distlib==0.3.4
distro-info===0.23ubuntu1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.3
pandas==1.4.1
parsimonious==0.8.1
platformdirs==2.5.1
protobuf==3.19.3
pycryptodome==3.12.0
PyGObject==3.36.0
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
requests-unixsocket==0.2.0
rlp==2.0.1
six==1.16.0
toolz==0.11.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
unattended-upgrades==0.1
urllib3==1.26.8
varint==1.0.2
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
yarl==1.7.2